requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
tensorflow>=2.8.0

# Web framework
flask>=2.0.0

# Visualization
plotly>=5.3.0
dash>=2.0.0

# Data processing
pyyaml>=6.0
tqdm>=4.62.0
nltk>=3.6.3

# NLP and Deep Learning
transformers>=4.11.0
sentence-transformers>=2.1.0
torch>=1.10.0
torchaudio>=0.10.0
torchvision>=0.11.0

# Audio processing
librosa>=0.9.0
soundfile>=0.10.0

# Image processing
Pillow>=9.0.0
opencv-python-headless>=4.5.0

# Performance monitoring
psutil>=5.9.0

# Testing and benchmarking
pytest>=6.2.5
pytest-benchmark>=3.4.1

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0
